Partially re-enable `rabbitmqadmin` HTTPS support for Python < 2.7.9
Prior to 2.7.9, Python didn't have a proper SSLContext and TLS
configurability of libraries such as httplib was limited to say the
least. To add insult to injury, said Python 2.x versions have SSLv3
enabled while RabbitMQ TCP listeners disable it in modern versions.

See the comments in https://github.com/rabbitmq/rabbitmq-management/issues/225 for more
information.

This can be worked around by switching to urllib2 but it's too risky
a change for 3.6.x in my opinion:
http://anotherlayer.net/2014/10/16/forcing-tls1-in-pythons-urllib2-on-osx/

Fixes #225 (to the extent we are willing to do this in 3.6.x and given
our team resources available).