Switch testsuite to common_test
At the same time, we drop the use of the SwiftMQ proprietary client
library in favor of two FOSS libraries: amqpnetlite for .NET and Qpid
Proton for Java.

Both libraries are used to implement the same tests, though only three
tests are currently implemented with Qpid Proton.

The "at most once" test case was not ported at all because none of those
libraries seem to support QoS.

[#121412149]