rabbit_mnesia_rename: Backup local tables only
The connection tracking tables are not replicated because the table
tracking connections on node A logically exists only the node A.

The backup made during the rename of a node failed because it wanted to
access a remote offline node to backup its connection tracking tables.
Obviously it didn't work. The solution is to not backup those tables.
This is correct because they are only relevant while the node is
running.

References #500.
[#116521809]