Remove all bits of the old test framework
Now that everything was moved to common_test, we don't need the old
framework anymore.

The only target still in use was `cluster-other-node`. As it was only
used by rabbitmq-java-client, the target was moved there.

The `test` target, which was an alias to `tests` (the regular Erlang.mk
target), was removed at the same time.

[#127356157]