Use rabbitmq_stomp instead of rabbitmq_metronome
The metronome plugin is an example and not supposed to be used
in production. It has certain assumptions that do not necessarily
hold when a test suite that messes with broker state is executed.

It therefore makes more sense to use rabbitmq_stomp or another
plugin that is commonly deployed in production systems.

Per discussion with @hairyhum.