rabbit_web_dispatch_SUITE: Start inets where we need it
Before this patch, inets was started during the suite setup on the
common_test control node, but we used it on the started broker. This
remained unnoticed while the plugin was based on webmachine and mochiweb
which started inets too.

After the move to Cowboy, inets is not started anymore on the broker so
we need to explicitely start it.