Choose master from intersection of sync slaves and policy nodes
Previously, when applying an HA policy with the 'nodes' type, if the
current master was not part of the new policy, then the first
synchronised slave was chosen as the new master. This would happen even
if the node was not part of the new policy.

For example, if we began with:

Master A
Slaves [B, C]

Then requested a new policy of just [C], we'd end up with:

Master B
Slaves [C]

Instead, choose the master from the intersection of (nodes in
the policy) and (synchronised slaves).

Issue #990
[#126767013]