rabbitmq_management_visualiser is gone for good
We cannot remove it in 3.6.x but it's really time for it to go.
It has been effectively unmaintained (and very much unused)
since at least 2012. In fact, it has always been experimental
and it wasn't a very good idea to include it into the distribution
in the first place.

We cannot continue carrying this dead weight forever.

Per discussion with other team members.