Makefile: Use $(PROJECT_VERSION) instead of $(VERSION)
This is the variable expected and used by Erlang.mk

When working from Git, the variable can be specified on the command
line. This is particularily useful when using `make run-broker` or `make
source-dist`.

When working from the source archive created by `make source-dist`, the
version is taken from `git-revisions.txt`, a file created during the
creation of the source archive. If the variable is specified on the
command line as well, it takes precedence over the one in
`git-revisions.txt`, like any Makefile variables.

[#130992027]