Move from .app.src to Makefile variables
This is the recommended way with Erlang.mk.

When working from the source archive created by `make source-dist`,
the version is taken from `git-revisions.txt`, a file created during
the creation of the source archive. When used for the broker, the
version is inherited from rabbitmq-serevr-release. If the variable is
specified on the command line as well, it takes precedence over the one
in `git-revisions.txt`, like any Makefile variables.

Erlang.mk automatically adds $(PROJECT)_sup to the registered processes.
We'll need to rename `amqp_sup` to that name or modify Erlang.mk to
avoid this bogus registered name.

[#130992027]