Move from .app.src to Makefile variables
This is the recommended way with Erlang.mk.

By default, the version is computed with git-describe(1) or inherited
from rabbitmq-server-release when creating the source archive.
One can override the version from the command line by setting the
`PROJECT_VERSION` variable.

One nice bonus is that `make run-broker` will present a meaningful
version by default, and one can set the version with `PROJECT_VERSION`.

[#130992027]