Support publish to Hex.pm
To do it, run:
    make hex-publish PROJECT_VERSION=3.6.6

You can also publish documentation to hexdocs.pm. The documentation is
generated first with the `docs` Erlang.mk target and the content of the
`doc` subdirectory is published. To do it, run:
    make hex-publish-docs PROJECT_VERSION=3.6.6

Generated files are published to Hex.pm so that users of this package
don't have to depend on rabbitmq-codegen and Python.

`git-revisions.txt` is generated in the process to carry the version of
the project. Not really standard, but it's the way other RabbitMQ
projects do it for now, so take advantage of the existing
infrastructure. Once a package is published to Hex.pm, it's good to
`make clean` so this file is removed.

Other RabbitMQ components can also be published to Hex.pm with a few
modifications. For them, rabbit_common/mk/rabbitmq-components.hexpm.mk
is sent in place of their copy of rabbitmq-components.mk. This modified
file pulls dependencies from Hex.pm using version $(PROJECT_VERSION),
instead of using Git.

All projects published to Hex.pm can use $(RABBITMQ_HEXPM_DEFAULT_FILES)
as the list of files to put in the `files` key in the .app file.

[#134268483]