rabbitmq-tools.mk: Add new target for Vim's Syntastic plugin
Those targets are used by both Erlang checkers to query compiler
options. Those options are required to properly report errors and
warnings.

This is a temporary commit to this project: once the patches to the
Syntastic Erlang checkers have stabilized, the change should be
submitted to Erlang.mk instead, and removed from this project.

[#136364285]