Support topic authorisation
Add all the resource options in the candidate variables for string substitutions.
This includes the routing key when publishing to an exchange topic. Let pass
when there is no match in the for clause for a topic resource (for backward
compatibility). This a best effort as some queries could fail when publishing
to a topic exchange, but we cannot know whether it's an omission of
the topic case in the query or the query actually blocks the topic.

Part of rabbitmq/rabbitmq-server#505