rabbit_ct_helpers: Handle Elixir modules in ensure_application_srcdir()
Elixir modules are placed in $SRCDIR/_build/.../ebin, not in
$SRCDIR/ebin. Handle that case when computing $SRCDIR from the path of a
module.

This fixes ensure_rabbitmqctl_app() the second time it is called and a
rabbitmqctl module was already loaded. The first time it was called,
everything worked because the application was not in the code path and
no module could be found by code:which().