Makefile: Use an intermediate file to list files to put in the source archive
Using xargs(1) is not always possible because it has a default limit on
the number of bytes it reads on stdin which can be too low for our need.
When we reach this limit, it breaks in a non-obvious way: the created
archive contains half of the expected files.

Now, we go through an intermediate file: the list is created once with
find(1) and sort(1) and consumed from tar(1) and zip(1).