packaging/debs: Simplify the Makefile wrapping dpkg-buildpackage(1)
The `check-changelog.sh` script is transformed to `update-changelog.sh`.
It doesn't care anymore if $UNOFFICIAL_RELEASE is set or not. It simply
verifies if the given version matches the last entry of
`debian/changelog` (without taking the package revision into account).
If the last entry doesn't match, it prepends a new automatically
generated entry.

Also, the Makefile takes both `VERSION` and `DEBIAN_VERSION` as input
variables. `VERSION` is the RabbitMQ version (eg. `3.6.7-alpha.62`)
and `DEBIAN_VERSION` is the same version formatted for Debian (eg.
`3.6.7~lpha.62`). If the latter is missing, it's computed from `VERSION`.

Regarding package signing, `GNUPG_PATH` is deprecated. The caller should
instead directly use `GNUPGHOME`, the documented variable expected by
GnuPG.

In the end, the Makefile is still compatible with the current way of
building a RabbitMQ release but it will simplify the job of Concourse.

As a bonus, there are two new scripts:
    o  scripts/get-debian-package-files-list.sh: List all files making a
       Debian package, based on a specified `.changes` file.
    o  scripts/compare-debian-versions.py: Compare two given versions
       (including the package revision) using APT. This helps when testing
       versioning.

[#131645209]