mk/rabbitmq-dist.mk: Build .ez archives for Elixir dependencies
We rely on the `mix_task_archive_deps` Mix task to create the project
.ez archive, as well as all its dependencies' archive. It also creates
an archive for Elixir runtime itself.

This task is also used to list the created archives, in order to filter
the one we want from the staled archives.

Write 'y' to mix(1) stdin, because it may ask confirmation to install
Hex.

References #rabbitmq/rabbitmq-auth-backend-uaa#3.
References #rabbitmq/rabbitmq-auth-backend-uaa#5.
[#136341089]