packaging/RPMS: Simplify the Makefile wrapping dpkg-buildpackage(1)
The RPM's changelog is now updated automatically if no entry match the
version being built. This is being handled by the parse-changelog.sh and
update-changelog.sh scripts.

Both VERSION and RPM_VERSION are still accepted. If the latter is
unspecified, it is derived from VERSION.

Package signing is now handled here, instead of the Umbrella. This makes
this Makefile on par with the Debian package's Makefile. This Makefile
still accepts KEYSDIR or GNUPG_PATH, but they are deprecated: one should
simply set GNUPGHOME.

In the end, the Makefile is still compatible with the current way of
building a RabbitMQ release but it will simplify the job of Concourse.

As a bonus, there is a new script called compare-rpm-versions.py: it
compares two given versions (including the package revision) using RPM.
This helps when testing versioning.

[#131645695]