Makefile: New `install` target to copy escripts
... to the directory specified with `$(PREFIX)`.

Handling of escripts in the `escript` is modified at the same time so
the result in this directory and in `$(PREFIX)` is the same.

The behavior is: we use hardlinks by default because they work
everywhere, including Windows. If `USE_SYMLINKS_IN_ESCRIPTS_DIR` is
specified, we use symlinks instead. However this is not the default
behavior because symlinks require privileges on Windows.

In the Makefile, I took this occasion to generalize a bit how we
distinguish actual escripts (rabbitmqctl) from linked escripts
(rabbitmq-plugins, rabbitmq-diagnostics). If we need to add more, it
should be easier now by editing the top of the Makefile. Management
of the `escript` directory and the install `$(PREFIX)` won't need
modifications.

[#138925175]