rabbit_mgmt_rabbitmqadmin_SUITE: Hard-code UTF-8-encoded queue name
... instead of using binary_to_list() on <<<"ööö"/utf8>>.

Erlang R16B03 took the "ö" individual bytes (from the two-byte encoded
UTF-8 character) and converted them to UTF-8. The queue was created with
the correct name because the correct UTF-8 string was passed to Python
and later to the management API. But because of the wrong conversion of
the expected queue name, the verification failed.