Makefile: Load rabbitmq-build.mk before erlang.mk
As an Erlang.mk plugin, rabbitmq-build.mk should be included in the
middle of erlang.mk but we can't do that here.

With the addition of `cth_fail_fast` to the TEST_DEPS, we need to move
the inclusion of rabbitmq-build.mk before erlang.mk. Otherwise, this new
dependency isn't fetched, but Erlang.mk still tries to build it:

    make[1]: Entering directory '.../rabbit_common'
    make[1]: *** .../cth_fail_fast: No such file or directory.  Stop.
    make[1]: Leaving directory '.../rabbit_common'
    erlang.mk:5161: recipe for target 'test-deps' failed
    make: *** [test-deps] Error 2

[#141155967]