Support typical RabbitMQ TCP socket options
The port option in the tcp_options entry takes now precedence
explicitly over the port option at the root of the env proplist.
It was previously the case, even if the opposite was advertized:
there would be 2 port entries in the Ranch options and Ranch
would take the one from tcp_options.

It's now possible to use strings for IP addresses, not only tuples.

The correct binding address and port are now used in the log message.

Fixes #69