system_SUITE: Accept `{error, closed}` as connection failures
With Erlang 18.3, there is a regression which causes the SSL connection
to crash with the following exception:

    ** {badarg,[{ets,update_counter,[1507362,#Ref<0.0.3.9>,-1],[]},
                {ssl_pkix_db,ref_count,3,...

When this exception reaches the connection process before the expected
TLS error, amqp_connection:start() returns `{error, closed}` instead.

Now, testcases tolerate `{error, closed}` as a return value: we know
that the connection was indeed terminated.

Signed-off: Gerhard Lazu <gerhard@rabbitmq.com>