packaging/standalone: Use an intermediate file to list files to archive
The problem is that xargs(1) has a limit on the number of bytes
it reads on stdin and we reach it. This leads to an archive
containing half of the expected files and xargs(1) never complains.
We apply the same fix which was previously appled in commit
b9dd874671814f5f67e19d280b6011e7080c8450.

[#141674113]