Register with peer discovery backend on start if the backend supports it
This introduces a new function to the peer discovery interface.
The idea is to let backends indicate if they support registration.
The goal is not to avoid calling the function, since Erlang
behaviours require all functions to be implemented anyway.
What we are trying to avoid with this is extra log entries
that say "[un]registering with ..." as they can be confusing
e.g. in single node installations.

References #1143.