Fix lager initialization code
Setting env var to 'undefined' is different from not setting it at
all, and lager is sensitive to this distinction in its 'log_root'
parameter. There will be a crash when rabbit is started without the
help of startup scripts (i.e. from `make shell` or in the embedded
mode).