Makefile: Fix malformed manpages
The generated manpage contains malformed .HP directives:
    .HP \w'\fBstop\fR\ 'u \fBstop\fR [\fIpid_file\fR]

The .HP directive takes one argument, the indentation. Thus, the first
line (which is not indented) should be on the next line:
    .HP \w'\fBstop\fR\ 'u
    \fBstop\fR [\fIpid_file\fR]

Manpages will be converted to mdoc(7) in master so we are sure of
the formatting. The DocBook versions will be dropped. This will also
simplify the build process.

References #1180.
[#143563295]