Regenerate Elixir-based .ez-s after changes
After changing any .erl-file and running `make dist` .ez files are
regenerated. But this is not the case for Elixir plugins - so it's
necessary to clean `plugins/` to refresh.

The difference is that for erlang.mk projects there is an explicit
dependencies on `ebin`, `include` and `priv` folder content -
https://github.com/rabbitmq/rabbitmq-common/blob/b2d635f5a27f8d4063950f9e8e93b8695c435ae0/mk/rabbitmq-dist.mk#L52

But there were no similar rule around
https://github.com/rabbitmq/rabbitmq-common/blob/b2d635f5a27f8d4063950f9e8e93b8695c435ae0/mk/rabbitmq-dist.mk#L91

So I've tried to add some reasonable dependencies there, and it helps.