scripts/rabbitmq-server: Work around signal handling issue with Dash
On Debian-like distributions, `/bin/sh` defaults to `/bin/dash` which
has a bug with signal handlers.

In the case of Dash, it looks like `set -e` (set at the beginning of
this script) gets precedence over signal handling. Therefore, when
`wait` is interrupted, its exit code is non-zero and because of `set
-e`, the script terminates immediately without running the signal
handler. To work around this issue, we use `|| true` to force that
statement to succeed and the signal handler to properly execute.

Replace the use of `-e` on the shebang line by a standalone `set -e`,
like other scripts. This way, the script behavior remains the same if
the script is started as an argument to a shell. For instance:
    bash ./rabbitmq-server

Bump the copyright year to 2017.

Signed-off-by: Patrick Sodré <sodre@sodre.co>

Fixes #1192.