Reload apps on config update.
Configuration update requires applications to be unloaded.
Some of the rabbit dependencies are loaded on VM boot, so they
should be unloaded before config update.
Applications which are already running during config update cannot be
unloaded (e.g. kernel) and warning should be reported that config
cannot be applied to these applications