Move rabbit_memory_monitor:memory_use/1 to vm_memory_monitor
The function is renamed to get_memory_use/1 in the process, to match
other functions in the vm_memory_monitor module.

The file_handle_cache uses that function. By moving it to
vm_memory_monitor, we remove a dependency of file_handle_cache on a
rabbit_* module. This will allow us to move both file_handle_cache and
vm_memory_monitor to rabbitmq-common and help remove a dependency in the
Erlang client on the broker because the former calls
file_handle_cache:obtain/0.

To sum up the situation before this commit with a graph:

  amqp_network_connection --[depends on]--> file_handle_cache
            (amqp_client)                   (rabbit)

        file_handle_cache --[depends on]--> rabbit_memory_monitor
                 (rabbit)                   (rabbit)

And the situation we want to reach after the move:

  amqp_network_connection --[depends on]--> file_handle_cache
            (amqp_client)                   (rabbit_common)

        file_handle_cache --[depends on]--> vm_memory_monitor
          (rabbit_common)                   (rabbit_common)

    rabbit_memory_monitor --[depends on]--> vm_memory_monitor
                 (rabbit)                   (rabbit_common)

[#118490793]