Makefile: Do not set -Werror when building src/*_compat.erl
Those modules might trigger warnings which can't be turned off. This
is the case with ssl:connection_info/1 which is removed in Erlang 20:
erl_lint emit a warning which can't be silenced. Therefore, -Werror
breaks the build for those modules.

The solution is to add those compat modules to `$(ERLC_EXCLUDE)` so that
Erlang.mk skips them when building the application.

We compile them ourselves without -Werror and patch the .app file to
append those compat modules to the list of modules.

[#145213041]