rabbitmq-test.mk: Add a *-on-concourse target
It allows to run any target on Concourse. All test dependencies are
taken locally and uploaded to Concourse before the make target is
executed. The test-output is written to logs/on-concourse.

By default, the oldest version of Erlang supported by the branch is
used. One can override that with the $(ERLANG_VERSION) make variable.

Here are some examples:

    gmake tests-on-concourse
    gmake dialyzer-on-concourse
    gmake ct-on-concourse V=1
    gmake ct-unit-on-concourse t=parallel_tests:pid_decompose_compose

[#146078591]