Make exchange consistent across restarts
The message hash used to be compared to a random number to choose the
destination. This random number would change when the broker restarts, making
the exchange not consistent across restarts. This commit replaces
the random number by a hash based on the source and the destination,
which is the same across restarts.

Fixes #32