mk/rabbitmq-early-test.mk: Do not set XREFR if target is `distclean{,-xrefr}`
Erlang.mk removes the copy of xrefr(1) during `make distclean-xref` (and
thus `make distclean`) because it assumes the script was fetched and can
be refetched.

In our case, the script is committed to this repository and all projects
are using it. Therefore, `make distclean` shouldn't remove our script.

To work around this, we do not set `$(XREFR)` if the target is either
`distclean` or `distclean-xref`.