mk/xrefr: Add Elixir-based applications to code path
For Erlang-based applications, the path is expected to be
`$application/ebin` and Erlang.mk sets up everything for us.

For Elixir-based applications, the path is something like
`$application/_build/$env/lib/*/ebin`. To add those, first we make sure
that $(ERLANG_MK_RECURSIVE_DEPS_LIST) is generated because it will
contains all build and runtime dependencies paths, and we pass that
filename to xrefr(1) using an environment variable.

Then xrefr(1) parses that file and looks at all of them if they contain
`_build/dev/lib/*/ebin` directories. If that's the case, those are added
to the code path.

Later in the script, the code path is added to xref library path.

[#140125673]