Ensure session process has time to finish processing on connections shutdown.
When an amqp 1.0 connection is closed very shortly after sending a large
batch of frames it is possible that the supervision tree terminates
before the session process has had time to process them all.
This change exposes a `call` function on the session that is called upon
the reader receiving an `v1_0.close{}` frame. This ensure the active sessions
get at least 5 seconds to empty their internal message buffers.