Remove hibernate from GM
We don't want to use the backoff/hibernate feature because we have
observed that the GM process is suspended half of the time.

We really wanted to replace gen_server2 with gen_server, but it was more
important to keep changes in 3.6 to a minimum. GM will eventually be
replaced, so switching it from gen_server2 to gen_server will be soon
redundant. We simply do not understand some of the gen_server2
trade-offs well enough to feel strongly about this change.

[#148892851]

Signed-off-by: Gerhard Lazu <gerhard@rabbitmq.com>