rabbitmqadmin: add --base-uri
When specified, hostname, port, username and password
are taken from the URI. URI path is intentionally ignored
because AMQP(S) URIs use path to configure target virtual host
but HTTP APIs typically use URI paths to specify an endpoint
prefix (--path-prefix in rabbitmqadmin).

To avoid confusion we log a warning and ask the user to
specify --vhost and/or --path-prefix explicitly.

Closes #437.