Do not crash command loading if unable to read enabled plugins file.
Some distributions set no permissions to read the enabled plugins
file location. This can cause a crash when running not as root or
rabbitmq user.
It was decided that plugins commands can be skipped when printing
the help message if the user have no access to the enabled plugins
file.
The warning message will be displayed in that case.

Addresses https://github.com/rabbitmq/rabbitmq-server/issues/1296
[#149079123]