Set queue state to 'stopped' when terminating.
If a queue process is stopped by a vhsot supervision tree, it should
be visible in the management UI and the `list_queues` command output.
Setting the queue state to `stopped` is easier to reason about than
checking the vhost aliveness status on a remote node.
If queue will be restarted or migrated to a different node, the status
will be set to `live`.
If a node was not stopped normally the queues will have `live`
state. If we cannot recover the queues thay should be marked
`stopped` to be reported to rabbitmqctl and management UI

Fixes #1303
[#148409695]