amqp_network_connection: Expect the socket_closed just after `Connection.Close`
... from the server.

This is fine because the server might not want to wait for all clients
to answer with `Connection.Close-Ok` and prefer to close the TCP
connection immediately instead.

Without this, the connection process exits with
`socket_closed_unexpectedly`.

This is a minor issue, but it was discovered while tracking the problem
fixed in commit b35dc5e74686a928f67cbd7f2611e5dd1e43ac2d. With those two
fixes, a server-initiated connection close is handled more gracefully.

[#149584741]