Restart stats timer for gen_server2 process.
Stats timer can be restarted after a stats message,
waking up grom hibernation and a message during backoff
period.
Only stats message should reset the old timer, because
stop_stats_timer resets valid timer and leaves it
if the timer already triggered, which means that
the message is in the process message box and will
be processed, and the timer will be reset.
After waking up and message during backoff period
the timer will be started only if it was successfully
stopped.

[#150544915]