Revert "packaging: Force GPG v3 signatures for the SLES 11 RPM package"
Using GPG v3/v4 signatures wasn't the problem. My test was wrong: the
tested package was not signed. So even a GPG v3 signature is refused by
RPM.

In fact, it looks like RPM 4.4 has issues with package signature
verifications. Perhaps related to the SHA1 digest (but using an MD4
digest doesn't solve the problem). The cause is unclear.

Only RPM 4.4 is available on SLES 11. So for now, revert this change.

This reverts commit dc783c737a637348ee01c2e33333c0339e30d3ba.