Revert "Ensure that exit code 0 is used for "stop" command"
This reverts commit 1518b601036351e3f8686bab9c8ae87a8d5af0a3.

There is no consensus on how it should work. In general it is impossible
to tell a stopped node from a node we could not contact (with some
specific exceptions reported by modern net_kernel version such as
the Erlang cookie mismatch), so some believe stop should be
idempotent and others believe it should not.