Bump rabbit.num_ssl_acceptors default to 10
Previously the value was limited to 1 due to a bug in Ranch 1.0:
it didn't depend on the ssl app, which means Ranch/ssl shut down order
was not guaranteed. This led to scary but entirely harmless exception traces
in the log, one per TLS acceptor.

The issue is no longer present in Ranch 1.3.x and 1.4.0.

Per discussion with @essen.