Make rabbit_networking:close_connection/2 more benign
The function relies on a fanout to all running cluster nodes
which in turn check their local connection process group.
All errors that might arise are ignored.

There are at least two scenarios where the current behavior makes
little sense:

 * A connection process has just terminated (but is still present in the stats DB)
 * A node could not respond (or respond in time) to the fanout

In both cases claiming that connection pid is "not a connection pid"
is misleading and likely will leave the user puzzled. This change
makes connection_close/2 more benign: it simply logs a warning and returns
instead of throwing.

There are only two call sites that use it: rabbitmqctl and an HTTP API
handler, so the chances of non-connection pids passed in are pretty slim.

Per discussion with @gerhard.