Travis CI: Run `make xref` in addition to `make tests`
While here, test on Erlang 19.3 (instead of 19.0) and Erlang 20.1. Also,
we stop testing on Erlang 17.5 and 18.3: it helps reduce the number of
jobs in Travis CI and allow it to go over all changes quicker.

Finally, the sections are reordered so the scripts are at the end. It
may help in the future if we want to template all the `.travis.tml`
files.

[#152509619]