rabbitmq-components.mk: Switch base branch to `v3.7.x`
Therefore, after we branch `v3.7.x`, rabbitmq-components.mk will try to
determine if a development branch was forked from `v3.7.x` (instead of
`stable`) and fallback to `master` if it's not the case.

While here, handle the case where the `master` branch doesn't exist
locally. This happens when using `git clone -b $branch` or in CI.