Makefile: Bundle Elixir dependencies in the RabbitMQ source archive
mix(1) doesn't seem to provide a nice way to provide dependencies'
sources ourselves. The "solution" is to override HOME and set it to
$(DEPS_DIR), call mix `deps.get` and remove the fetched dependencies.
The goal is to populate the Hex cache and bundle it in the source
archive. No worries, you can thank me later.

Later, when we build the source archive, we override HOME again to point
mix(1) to the Hex cache: it is now happy to "fetch" dependencies. This
part is managed by `rabbitmq-mix.mk` which is copied `rabbitmq-mix.mk`
from rabbit_common to the Elixir components (e.g. rabbitmq_cli) in the
source archive.

References rabbitmq/rabbitmq-server-release#61.
[#153358632]