Support kernel.net_ticktime in Cuttlefish configuration
net_ticktime, if present in the .conf file, is set up with
net_kernel:set_net_ticktime/1 before the configuration
of the .conf is applied to the other applications. The kernel
is already running when the configuration is applied, so net_ticktime
is set in a specific way.

[#155393098]

Fixes #1522

(cherry picked from commit fdb3e7528431491e5d8718ef65e28c82efeb898b)