Split single Mnesia transaction that runs on_node_down
Rather than using a single Mnesia transaction to clean queues when a
node goes down, split it into many transactions so that other Mnesia
operations can make progress. This is especially important when nodes
join the cluster, since Mnesia on the newly started node will not be
able to synchronise if there is a long-running Mnesia transaction.

This commit is not complete, we need feedback on the comments left in
the code before we can settle on a final version that can be merged.

For more context, see #1513

Partner-in-crime: @essen