Make maximum number of Erlang atoms configurable via env
This was the only property in SERVER_ERL_ARGS that was not configurable.
Even though it's unlikely that this will need to be increased above the
existing 5 million default, we are consistent in allowing all properties
in SERVER_ERL_ARGS to be cofigured.

(cherry picked from commit b1e163fd6f385af03c65dbe421def3597ce01df2)