Take all hostnames from an `instancesSet`
Before this patch, the code was only taking the first item in an
`instancesSet`, using `proplists:get_value()`. Therefore it would ignore
all other entries in that list. Unfortunately, if that first host was
the VM running the code itself, the plugin would find no RabbitMQ nodes
to cluster with.

The patch simply consists of going through all `item` entries in that
`instancesSet` list.

Fixes #20.
[#153749132]