Accept -t/--timeout before command name [again]
There are existing users (including RabbitMQ core tests) that
use that sequence because it was previously in the docs.

This relaxes command validation so that if a command-specific
flag overrides a global, we accept it as long as their types match.

Per discussion with @hairyhum.

(cherry picked from commit 56991de0f3c0ce68bcffe4c9e3b51eec87d5b3b8)