Force-delete queues, which have no live master or slave processes.
Fixes #1501
[#155801556]
If a queue is configured to not be promoted (via ha-promote-on-shutdown: when-synced)
queue.delete can hang. Make it check for process existense first and
force-delete if no master of slave processes are running.
Do not force-delete if if_empty is set, since there is no
way to check that the queue is empty.

(cherry picked from commit 3e7bd564bda36c1bbb9e3b59b61509d0982a88ec)