Makefile: Move `--no-recursion` before `-T`
On tar(1) command line, `--recursion` and `--no-recursion` are
applicable to directories/files/files list following them, not before
them. Therefore we need to specify the files list (`-T`) after
`--no-recursion`.

Before this change, the default mode was used for everything in the
files list, which is `--recursion`. And because the files list was
listing all directories and files, files were stored multiple times in
the resulting tar archive: once for their entry in the files list and
once for each parent directories.

This fix greatly reduce the size of the source archive: it goes from 7.1
MiB to 2.4 MiB for the xz-compressed tar.