Do not try to stop all slaves from slave process. Stop it and let another be promoted
There can be a race condition when a master queue is briefly restarted.
If master rejoins a stopping GM it is also stopped.
In that case it sould be sefier to stop the slave and let another be
promoted and also stopped. If master return it will either rejoin
slaves or create a new GM.