rabbit_mgmt_format: Convert `sni_hosts` to something compatible with jsx
jsx expects a proplist key to be an Erlang binary, an integer or an
atom. It doesn't accept an Erlang string, probably because it's
difficult to distinguish from a list of integers.

Therefore the patch consists of converting all hostnames under
`sni_hosts` from strings to binaries.

Fixes rabbitmq/rabbitmq-management#567.