Revert "Take policy-configured max-priority into account"
This reverts commit f5aa1fbe043395806d9b9ed8780892924431466c.

This feature wasn't available in the original implementation for a reason:
policies are dynamic and can change after a queue's been declared. However,
queue priorities are (at least currently) set in stone from the moment of
queue creation. This was mentioned in the docs but not explicitly enough and got overlooked.

Credit for the [re-]discovery goes to @acogoluegnes :)

References #1590.

[#157380396]